使用語意標籤 & 重複登出報錯問題解決

+++ src/app/Http/Controllers/LoginController.php
@@ -20,17 +20,16 @@ class LoginController extends Controller
             $token = User::query()->where('id', '=', Auth::id())->get()[0]->createToken('x-accessToken');
             return redirect()->route('dashboard')->cookie('x-access-token', $token->plainTextToken);
         }
-
-
         return back()->withErrors([
-//            'student_id' => 'The provided credentials do not match our records.',
             'cantFind' => '帳號或密碼錯誤',
         ])->onlyInput('student_id');
     }
 
     public function logout(Request $request)
     {
-        $request->user()->tokens()->delete();
+        if ($request->user() != null) {
+            $request->user()->tokens()->delete();
+        }
         Auth::logout();
 
         $request->session()->invalidate();
@@ -39,25 +38,6 @@ class LoginController extends Controller
         return redirect('/');
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -67,52 +47,7 @@ class LoginController extends Controller
      */
     public function store(Request $request)
     {
-        dd($request->all());
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
+//        dd($request->all());
         //
     }
 }
